# app/Models/Passenger.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Passenger extends Model
{
    use HasFactory;

    protected $fillable = ['name', 'idNumber', 'phoneNumber', 'itinerary_id'];

    public function itinerary() {
        return $this->belongsTo(Itinerary::class);
    }
}