# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Vuelos",
        "description": "Documentación de la API para gestionar vuelos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/passengers": {
            "get": {
                "tags": [
                    "Passengers"
                ],
                "summary": "Obtener todos los pasajeros",
                "operationId": "394569dcd7779a3b1f56d881fa221a02",
                "responses": {
                    "200": {
                        "description": "Lista de pasajeros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Passenger"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error fetching passengers"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Passengers"
                ],
                "summary": "Crear un pasajero",
                "operationId": "f4c846bd1a08e31831bff994899aa418",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "idNumber": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "phoneNumber": {
                                        "type": "string",
                                        "example": "555-1234"
                                    },
                                    "itinerary_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pasajero creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al guardar el pasajero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error saving passenger"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/passengers/{id}": {
            "get": {
                "tags": [
                    "Passengers"
                ],
                "summary": "Obtener un pasajero por ID",
                "operationId": "3f64e30d6bff7658877062761b4ade90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pasajero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del pasajero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Passenger"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pasajero no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Passenger not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error fetching passenger"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Passengers"
                ],
                "summary": "Actualizar un pasajero",
                "operationId": "05bf07ab27b1163f073ac785bde9687a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pasajero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "idNumber": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "phoneNumber": {
                                        "type": "string",
                                        "example": "555-1234"
                                    },
                                    "itinerary_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pasajero actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Passenger updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pasajero no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Passenger not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error updating passenger"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Passengers"
                ],
                "summary": "Eliminar un pasajero",
                "operationId": "439cc243b528531bb91b9422d2c0ae8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pasajero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pasajero eliminado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Passenger deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pasajero no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Passenger not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error deleting passenger"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Passenger": {
                "title": "Passenger Schema",
                "description": "Esquema de un pasajero",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "idNumber": {
                        "type": "string",
                        "example": "123456789"
                    },
                    "phoneNumber": {
                        "type": "string",
                        "example": "555-1234"
                    },
                    "itinerary_id": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            }
        }
    },
...
}